feat: [유경우] 사용자가 카카오 결제를 요청하면 카카오에서 결제 페이지를 보내도록 합니다 [SDY-85]

+++ src/main/java/com/dyes/backend/domain/order/service/OrderServiceImpl.java
@@ -6,28 +6,29 @@ import com.dyes.backend.domain.cart.entity.ContainProductOption;
 import com.dyes.backend.domain.cart.repository.CartRepository;
 import com.dyes.backend.domain.cart.repository.ContainProductOptionRepository;
 import com.dyes.backend.domain.cart.service.CartService;
+import com.dyes.backend.domain.order.controller.form.KakaoPaymentApprovalRequestForm;
+import com.dyes.backend.domain.order.controller.form.KakaoPaymentRejectRequestForm;
+import com.dyes.backend.domain.order.controller.form.OrderConfirmRequestForm;
+import com.dyes.backend.domain.order.controller.form.OrderProductRequestForm;
 import com.dyes.backend.domain.delivery.entity.Delivery;
 import com.dyes.backend.domain.delivery.repository.DeliveryRepository;
-import com.dyes.backend.domain.order.controller.form.OrderConfirmRequestForm;
-import com.dyes.backend.domain.order.controller.form.OrderProductInCartRequestForm;
-import com.dyes.backend.domain.order.controller.form.OrderProductInProductPageRequestForm;
 import com.dyes.backend.domain.delivery.entity.DeliveryStatus;
-import com.dyes.backend.domain.order.entity.OrderedProduct;
-import com.dyes.backend.domain.order.entity.OrderedPurchaserProfile;
-import com.dyes.backend.domain.order.entity.ProductOrder;
+import com.dyes.backend.domain.order.entity.*;
 import com.dyes.backend.domain.order.repository.OrderRepository;
 import com.dyes.backend.domain.order.repository.OrderedProductRepository;
 import com.dyes.backend.domain.order.repository.OrderedPurchaserProfileRepository;
 import com.dyes.backend.domain.order.service.admin.response.OrderDetailInfoResponse;
 import com.dyes.backend.domain.order.service.admin.response.OrderProductListResponse;
 import com.dyes.backend.domain.order.service.admin.response.OrderUserInfoResponse;
-import com.dyes.backend.domain.order.service.user.response.form.OrderListResponseFormForUser;
-import com.dyes.backend.domain.order.service.user.request.OrderConfirmRequest;
-import com.dyes.backend.domain.order.service.user.request.OrderedProductOptionRequest;
-import com.dyes.backend.domain.order.service.user.request.OrderedPurchaserProfileRequest;
-import com.dyes.backend.domain.order.service.user.response.*;
-import com.dyes.backend.domain.order.service.user.response.form.OrderConfirmResponseFormForUser;
 import com.dyes.backend.domain.order.service.admin.response.form.OrderListResponseFormForAdmin;
+import com.dyes.backend.domain.order.service.user.request.*;
+import com.dyes.backend.domain.order.service.user.response.OrderConfirmProductResponse;
+import com.dyes.backend.domain.order.service.user.response.OrderConfirmUserResponse;
+import com.dyes.backend.domain.order.service.user.response.OrderOptionListResponse;
+import com.dyes.backend.domain.order.service.user.response.form.OrderConfirmResponseFormForUser;
+import com.dyes.backend.domain.order.service.user.response.form.OrderListResponseFormForUser;
+import com.dyes.backend.domain.payment.service.PaymentService;
+import com.dyes.backend.domain.payment.service.request.KakaoPaymentApprovalRequest;
 import com.dyes.backend.domain.product.entity.Product;
 import com.dyes.backend.domain.product.entity.ProductMainImage;
 import com.dyes.backend.domain.product.entity.ProductOption;
@@ -37,12 +38,17 @@ import com.dyes.backend.domain.user.entity.Address;
 import com.dyes.backend.domain.user.entity.User;
 import com.dyes.backend.domain.user.entity.UserProfile;
 import com.dyes.backend.domain.user.repository.UserProfileRepository;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.web.servlet.view.RedirectView;
 
 import java.time.LocalDate;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
 import static com.dyes.backend.domain.delivery.entity.DeliveryStatus.PREPARING;
 
@@ -60,14 +66,38 @@ public class OrderServiceImpl implements OrderService {
     final private OrderedPurchaserProfileRepository orderedPurchaserProfileRepository;
     final private DeliveryRepository deliveryRepository;
     final private CartService cartService;
+    final private PaymentService paymentService;
     final private AuthenticationService authenticationService;
 
-    // 장바구니에서 상품 주문
-    @Override
-    public boolean orderProductInCart(OrderProductInCartRequestForm requestForm) {
+    public RedirectView purchaseReadyWithKakao(OrderProductRequestForm requestForm) throws JsonProcessingException {
+        log.info("purchaseKakao start");
+
+        OrderProductRequest request = new OrderProductRequest(
+                requestForm.getUserToken(),
+                requestForm.getOrderedPurchaserProfileRequest(),
+                requestForm.getOrderedProductOptionRequestList(),
+                requestForm.getTotalAmount(),
+                requestForm.getFrom()
+                );
+        RedirectView redirectUrl = paymentService.paymentTemporaryDataSaveAndReturnRedirectView(request);
+
+        return redirectUrl;
+    }
+    public boolean approvalPurchaseWithKakao (KakaoPaymentApprovalRequestForm requestForm) throws JsonProcessingException {
+        log.info("approvalPurchaseKakao start");
+        KakaoPaymentApprovalRequest request = new KakaoPaymentApprovalRequest(requestForm.getUserToken(), requestForm.getPg_token());
+        boolean result = paymentService.paymentApprovalRequest(request);
+        log.info("approvalPurchaseKakao end");
+        return result;
+    }
+    public boolean rejectPurchaseWithKakao (KakaoPaymentRejectRequestForm requestForm) {
+        boolean result = paymentService.paymentRejectWithKakao(requestForm);
+        return result;
+    }
+    // 상품 주문
+    public boolean orderProductInCart(OrderProductRequestForm requestForm) {
         log.info("orderProductInCart start");
         try {
-
             OrderedPurchaserProfileRequest profileRequest = OrderedPurchaserProfileRequest.builder()
                     .orderedPurchaserName(requestForm.getOrderedPurchaserProfileRequest().getOrderedPurchaserName())
                     .orderedPurchaserContactNumber(requestForm.getOrderedPurchaserProfileRequest().getOrderedPurchaserContactNumber())
@@ -76,27 +106,31 @@ public class OrderServiceImpl implements OrderService {
                     .orderedPurchaserZipCode(requestForm.getOrderedPurchaserProfileRequest().getOrderedPurchaserZipCode())
                     .orderedPurchaserAddressDetail(requestForm.getOrderedPurchaserProfileRequest().getOrderedPurchaserAddressDetail())
                     .build();
-            final String paymentNumber = "랜덤한 결제 정보" + UUID.randomUUID();
+
             final String userToken = requestForm.getUserToken();
             final int totalAmount = requestForm.getTotalAmount();
             User user = authenticationService.findUserByUserToken(userToken);
+            if (user == null) {
+                return false;
+            }
             List<OrderedProductOptionRequest> orderedProductOptionRequestList = requestForm.getOrderedProductOptionRequestList();
 
-            saveOrderedData(profileRequest, paymentNumber, totalAmount, user, orderedProductOptionRequestList);
-
-            // 주문한 상품이 장바구니에 있으면 장바구니에서 목록 제거
+            saveOrderedData(profileRequest, totalAmount, user, orderedProductOptionRequestList);
 
             // 유저 토큰으로 장바구니 찾기
             Cart cart = cartService.cartCheckFromUserToken(userToken);
 
-            // 장바구니에 담긴 상품 리스트 불러오기
-            List<ContainProductOption> productOptionList = containProductOptionRepository.findAllByCart(cart);
+            // 주문한 상품이 장바구니에 있으면 장바구니에서 목록 제거
+            if (requestForm.getFrom().equals("cart")) {
+                // 장바구니에 담긴 상품 리스트 불러오기
+                List<ContainProductOption> productOptionList = containProductOptionRepository.findAllByCart(cart);
 
-            for (ContainProductOption containProductOption : productOptionList) {
-                for (OrderedProductOptionRequest orderedProductOptionRequest : orderedProductOptionRequestList)
-                    if (Objects.equals(containProductOption.getOptionId(), orderedProductOptionRequest.getProductOptionId())) {
-                        containProductOptionRepository.delete(containProductOption);
-                    }
+                for (ContainProductOption containProductOption : productOptionList) {
+                    for (OrderedProductOptionRequest orderedProductOptionRequest : orderedProductOptionRequestList)
+                        if (Objects.equals(containProductOption.getOptionId(), orderedProductOptionRequest.getProductOptionId())) {
+                            containProductOptionRepository.delete(containProductOption);
+                        }
+                }
             }
             return true;
         } catch (Exception e) {
@@ -104,47 +138,19 @@ public class OrderServiceImpl implements OrderService {
             return false;
         }
     }
-
-    // 제품 페이지에서 상품 주문
-    @Override
-    public boolean orderProductInProductPage(OrderProductInProductPageRequestForm requestForm) {
-        log.info("orderProductInProductPage start");
-
-        try {
-
-            OrderedPurchaserProfileRequest profileRequest = OrderedPurchaserProfileRequest.builder()
-                    .orderedPurchaserName(requestForm.getOrderedPurchaserProfileRequest().getOrderedPurchaserName())
-                    .orderedPurchaserContactNumber(requestForm.getOrderedPurchaserProfileRequest().getOrderedPurchaserContactNumber())
-                    .orderedPurchaserEmail(requestForm.getOrderedPurchaserProfileRequest().getOrderedPurchaserEmail())
-                    .orderedPurchaserAddress(requestForm.getOrderedPurchaserProfileRequest().getOrderedPurchaserAddress())
-                    .orderedPurchaserZipCode(requestForm.getOrderedPurchaserProfileRequest().getOrderedPurchaserZipCode())
-                    .orderedPurchaserAddressDetail(requestForm.getOrderedPurchaserProfileRequest().getOrderedPurchaserAddressDetail())
-                    .build();
-            final String paymentNumber = "랜덤한 결제 정보" + UUID.randomUUID();
-            final String userToken = requestForm.getUserToken();
-            final int totalAmount = requestForm.getTotalAmount();
-            User user = authenticationService.findUserByUserToken(userToken);
-            List<OrderedProductOptionRequest> orderedProductOptionRequestList = requestForm.getOrderedProductOptionRequestList();
-
-            saveOrderedData(profileRequest, paymentNumber, totalAmount, user, orderedProductOptionRequestList);
-
-            log.info("orderProductInProductPage end");
-            return true;
-        } catch (Exception e) {
-            log.error("Error occurred while ordering products in cart", e);
-            return false;
-        }
-    }
-
     // 상품을 주문하기 전에 확인하기
     @Override
     public OrderConfirmResponseFormForUser orderConfirm(OrderConfirmRequestForm requestForm) {
         try {
             OrderConfirmRequest request = new OrderConfirmRequest(requestForm.getUserToken());
+            List<OrderConfirmProductRequest> requestList = requestForm.getRequestList();
 
             final String userToken = request.getUserToken();
             // 유저 정보 찾기
             User user = authenticationService.findUserByUserToken(userToken);
+            if (user == null) {
+                return null;
+            }
             UserProfile userProfile = userProfileRepository.findByUser(user).get();
 
             // 반환될 유저 정보
@@ -162,22 +168,9 @@ public class OrderServiceImpl implements OrderService {
                 log.error("please set your user profile");
             }
 
-            // 유저 토큰으로 장바구니 찾기
-            Cart cart = cartRepository.findByUser(user).get();
-            log.info("cart: " + cart.getId());
-
-            // 장바구니에 담긴 상품 리스트 불러오기
-            List<ContainProductOption> productOptionList = containProductOptionRepository.findAllByCart(cart);
-            log.info("productOptionList: " + productOptionList.get(0).getProductName());
-
-            // 장바구니에 담긴 물건이 없으면 에러
-            if (productOptionList.size() == 0) {
-                throw new IllegalArgumentException("No exist product in the cart");
-            }
-            // 장바구니에 담긴 물건을 모조리 불러오기
             List<OrderConfirmProductResponse> productResponseList = new ArrayList<>();
-            for (ContainProductOption containProductOption : productOptionList) {
-                ProductOption productOption = productOptionRepository.findByIdWithProduct(containProductOption.getOptionId()).get();
+            for (OrderConfirmProductRequest productRequest : requestList) {
+                ProductOption productOption = productOptionRepository.findByIdWithProduct(productRequest.getProductOptionId()).get();
                 log.info("productOption: " + productOption.getOptionName());
 
                 ProductMainImage mainImage = productMainImageRepository.findByProductId(productOption.getProduct().getId()).get();
@@ -188,7 +181,6 @@ public class OrderServiceImpl implements OrderService {
                         .productName(productOption.getProduct().getProductName())
                         .optionId(productOption.getId())
                         .optionPrice(productOption.getOptionPrice())
-                        .optionCount(containProductOption.getOptionCount())
                         .productMainImage(mainImage.getMainImg())
                         .value(productOption.getAmount().getValue())
                         .unit(productOption.getAmount().getUnit())
@@ -234,7 +226,7 @@ public class OrderServiceImpl implements OrderService {
 
             // 주문한 상품 및 옵션 정보 가져오기
             Long totalPrice = 0L;
-            String productOrderId = order.getId();
+            Long productOrderId = order.getId();
             Delivery delivery = order.getDelivery();
             DeliveryStatus deliveryStatus = delivery.getDeliveryStatus();
             LocalDate orderedTime = order.getOrderedTime();
@@ -287,6 +279,9 @@ public class OrderServiceImpl implements OrderService {
     public List<OrderListResponseFormForUser> getMyOrderListForUser(String userToken) {
 
         User user = authenticationService.findUserByUserToken(userToken);
+        if (user == null) {
+            return null;
+        }
 
         // 모든 주문 내역 가져오기
         List<ProductOrder> orderList = orderRepository.findAllByUserWithUserAndDelivery(user);
@@ -297,7 +292,7 @@ public class OrderServiceImpl implements OrderService {
 
             // 주문한 상품 및 옵션 정보 가져오기
             Long totalPrice = 0L;
-            String productOrderId = order.getId();
+            Long productOrderId = order.getId();
             Delivery delivery = order.getDelivery();
             DeliveryStatus deliveryStatus = delivery.getDeliveryStatus();
             LocalDate orderedTime = order.getOrderedTime();
@@ -346,7 +341,7 @@ public class OrderServiceImpl implements OrderService {
     }
 
     // 주문 진행
-    public void saveOrderedData(OrderedPurchaserProfileRequest profileRequest, String paymentNumber,
+    public void saveOrderedData(OrderedPurchaserProfileRequest profileRequest,
                                 int totalAmount, User user, List<OrderedProductOptionRequest> orderedProductOptionRequestList) {
 
         final String purchaserName = profileRequest.getOrderedPurchaserName();
@@ -363,9 +358,9 @@ public class OrderServiceImpl implements OrderService {
         deliveryRepository.save(delivery);
 
         ProductOrder order = ProductOrder.builder()
-                .id(paymentNumber)
                 .user(user)
-                .totalAmount(totalAmount)
+                .orderStatus(OrderStatus.SUCCESS_PAYMENT)
+                .amount(new OrderAmount(totalAmount, 0))
                 .orderedTime(LocalDate.now())
                 .delivery(delivery)
                 .build();
